loghelper/requirements.txt
azure-eventhub
azure-identity
opentelemetry-api
azure-monitor-opentelemetry-exporter